+++ src/seeds/inventarioSeed.js
@@ -1,23 +1,27 @@
-import fakerbr    from 'faker-br';
+import fakerbr from 'faker-br';
 import Inventario from '../models/Inventario.js';
-import Campus     from '../models/Campus.js';
-//import DbConnect  from '../config/dbConnect.js';
+import Campus from '../models/Campus.js';
 
 export default async function inventarioSeed() {
-    //Coleta os campus cadastrados no banco de dados
+    // Coleta os campus cadastrados no banco de dados
     const campusList = await Campus.find({});
 
-    //Deleta todos os inventários existentes no banco de dados
+    if (campusList.length === 0) {
+        console.log("Nenhum campus encontrado. Execute campusSeed primeiro.");
+        return;
+    }
+
+    // Deleta todos os inventários existentes no banco de dados
     await Inventario.deleteMany({});
 
-    //Gera 50 inventários
+    // Gera 20 inventários
     for (let i = 0; i < 20; i++) {
         const randomCampus = campusList[Math.floor(Math.random() * campusList.length)];
 
         const inventario = {
             campus: randomCampus._id,
-            nome:   fakerbr.lorem.word(10),
-            data:   fakerbr.date.past(),
+            nome: fakerbr.lorem.word(10),
+            data: fakerbr.date.past(),
             status: fakerbr.random.boolean(),
         };
